Simulate the fishery dynamics for a short-lived semelparous species

Simulate the fishery dynamics for a short-lived semelparous species

Simulate(
  LifeHistory = NULL,
  Exploitation = NULL,
  nsim = 100,
  seed = 101,
  currentYr = format(Sys.Date(), "%Y"),
  silent = FALSE
)

Arguments

LifeHistory

A named list object with the life-history parameters, or the file path to a correctly formatted CSV file containing the LifeHistory (and optionally Exploitation) parameters. The LifeHistory object can optionally include the Exploitation parameters as well.

Exploitation

An named list object with the exploitation parameters or the file path to a correctly formatted Exploitation CSV file. Not needed if LifeHistory already includes the Exploitation parameters.

nsim

The number of stochastic simulations. Default is 100.

seed

The seed for the random number generator

currentYr

The current (most recent) year for the simulations

silent

Hide messages?

Value

A named list of class Simulation containing:

LifeHistory: The LifeHistory Object used for the simulations

Exploitation: The Exploitation Object used for the simulations, with the addition of Sel_at_Age (selectivity schedule)

At_Age_Time_Series: A data frame with the following columns

Sim: The simulation number

Age: The age class

Year: The year

Month: Abbreviated calendar month

Month_ind: Numerical index for the monthly timestep

N_fished: The number of fished animals in each age class for each timestep

N_unfished: The number of unfished animals in each age class for each timestep

N_unfished_eq: The equilibrium number of unfished animals in each age class for each timestep

B_fished: The biomass of fished animals in each age class for each timestep

B_unfished: The biomass of unfished animals in each age class for each timestep

SB_fished: The spawning biomass of fished animals in each age class for each timestep

SB_unfished: The spawning biomass of unfished animals in each age class for each timestep

Catch: The catch biomass in each age class for each timestep

Catch_n: The catch numbers in each age class for each timestep

Time_Series: A data frame with the following columns

Sim: The simulation number

Year: The year

Month: Abbreviated calendar month

Month_ind: Numerical index for the monthly timestep

N_fished: The total number of fished animals in each timestep

N_unfished: The total number of unfished animals in each timestep

B_fished: The biomass of fished animals in each timestep

B_unfished: The biomass of unfished animals in each timestep

SB_fished: The spawning biomass of fished animals in each timestep

SB_unfished: The spawning biomass of unfished animals in each timestep

Catch: The catch biomass in each timestep

Recruits: The number of recruits in each timestep

SPR: The equilibrium spawning potential ratio in each timestep

Effort: The fishing effort in each timestep

F_mort: The fishing mortality in each timestep

Rec_Devs: The recruitment deviation in each timestep

Details

See the example SLAM::LifeHistory and SLAM::Exploitation objects and help documentation (?LifeHistory and ?Exploitation) for examples.