options {slam}	R Documentation

Options for the ‘slam’ package

Description

Function for getting and setting options for the slam package.

Usage

slam_options(option, value)

Arguments

option
character string indicating the option to get or set (see details). If missing, all options are returned as a list.

value
Value to be set. If omitted, the current value is returned.

Details

Currently, the following options are available:

"max_dense":

numeric specifying the maximum length of dense vectors (default: 2^24).

See Also

simple_sparse_array

Examples

## save defaults
.slam_options <- slam_options()
.slam_options

slam_options("max_dense", 2^25)
slam_options("max_dense")

## reset
slam_options("max_dense", .slam_options$max_dense)

[Package slam version 0.1-46 Index]